services/user/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv==1.0.1
passlib[bcrypt]
pg8000